third_party/blink/web_tests/dom/mutation-event-tests/fast/forms/select/select-set-length-with-mutation-expected.txt

CONSOLE WARNING: Listener added for a synchronous 'DOMNodeRemoved' DOM Mutation Event. This event type is deprecated (https://w3c.github.io/uievents/#legacy-event-types) and work is underway to remove it from this browser. Usage of this event listener will cause performance issues today, and represents a risk of future incompatibility. Consider using MutationObserver instead.
CONSOLE WARNING: Listener added for a synchronous 'DOMNodeInserted' DOM Mutation Event. This event type is deprecated (https://w3c.github.io/uievents/#legacy-event-types) and work is underway to remove it from this browser. Usage of this event listener will cause performance issues today, and represents a risk of future incompatibility. Consider using MutationObserver instead.
Tests setting the .length of an HTMLSelectElement with mutation listeners registered.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS sel.length is 200
PASS sel.length is 100
PASS sel.length is 180
PASS successfullyParsed is true

TEST COMPLETE
